+++ app/src/main/kotlin/com/looker/kenko/ui/planEdit/PlanEditViewModel.kt
@@ -39,7 +39,7 @@ class PlanEditViewModel @Inject constructor(
 
     private val routeData: PlanEditRoute = savedStateHandle.toRoute<PlanEditRoute>()
 
-    private val planId: Long? = routeData.id.takeIf { it != -1L }
+    private val planId: Int? = routeData.id.takeIf { it != -1 }
 
     private val planStream: Flow<Plan?> = repo.get(planId)
